refactor transport and security to common
