+++ docs/plugins/common/security.md
@@ -1,0 +1,7 @@
+### Security
+| Variable Name | Type | Required | Default | Description |
+|---|---|---|---|---|
+| self_hostname | string | Yes | - | Hostname<br> |
+| shared_key | string | Yes | - | Shared key for authentication.<br> |
+| user_auth | bool | No | - | If true, use user based authentication.<br> |
+| allow_anonymous_source | bool | No | - | Allow anonymous source. <client> sections are required if disabled.<br> |
